refactor: 컴퓨터, 사용자 입력값 기능 모듈화
